fix(mention): 修復 @mention 通知未觸發 — 抽取共用 hook、補齊前端解析與後端編輯差異比對
## Why is this necessary?

- social/comment-input.tsx 缺少 @mention 解析 UI，mentionedUserIds 不會被送到 API，導致通知流程未觸發
- MentionInput 與 mention 追蹤邏輯重複嵌在 comment-section.tsx，無法被其他元件共用
- updateComment 未處理 mentionedUserIds，編輯留言新增 @mention 時不會發送 P1 通知

## How does it address?

- 建立 @daodao/features-mention 套件，將 useMentionInput() hook 與 MentionInput 元件抽出，供各留言入口統一引用
- comment-section.tsx 改用抽出的 hook，移除 inline 的 mentionedIds 狀態管理
- social/comment-input.tsx 整合 MentionInput 與 useMentionInput()，onSubmit 新增 mentionedUserIds 參數
- 新增 7 個 vitest 單元測試（getActiveMentionIds word-boundary 邏輯）
