+++ apps/product/package.json
@@ -22,6 +22,7 @@
     "@daodao/api": "workspace:*",
     "@daodao/assets": "workspace:*",
     "@daodao/auth": "workspace:*",
+    "@daodao/features-mention": "workspace:*",
     "@daodao/features-quiz": "workspace:*",
     "@daodao/i18n": "workspace:*",
     "@daodao/shared": "workspace:*",
