+++ apps/product/src/components/social/comment-input.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { MentionInput, useMentionInput } from "@daodao/features-mention";
+import type { MentionCandidate } from "@daodao/features-mention";
 import { Button } from "@daodao/ui/components/button";
 import { cn } from "@daodao/ui/lib/utils";
 import { Send } from "lucide-react";
@@ -13,9 +15,15 @@ import { REACTION_CONFIG, type ReactionTypeType } from "@/constants/reaction-typ
 export interface CommentInputProps {
   /** 目前選取的反應類型，決定 placeholder 文字 */
   reactionType?: ReactionTypeType | null;
-  onSubmit: (content: string, reactionType?: ReactionTypeType | null) => void;
+  onSubmit: (
+    content: string,
+    reactionType?: ReactionTypeType | null,
+    mentionedUserIds?: number[]
+  ) => void;
   disabled?: boolean;
   className?: string;
+  /** 可供 @mention 的使用者候選清單（由父元件從留言清單中衍生） */
+  participants?: MentionCandidate[];
 }
 
 // ============================================================================
@@ -23,14 +31,25 @@ export interface CommentInputProps {
 // ============================================================================
 
 /**
- * 留言輸入框
+ * 留言輸入框（含 @mention 支援）
  * - 根據 reactionType 動態切換 placeholder 引導文字
  * - 當 reactionType 改變時自動聚焦
+ * - 送出時傳送 mentionedUserIds 給父元件
+ *
+ * Task 2.1 調查：CommentInput 目前由 social/index.ts 匯出但尚未在任何頁面使用，
+ * 適合加入 @mention UI（所有 target type 均為具名用戶，不涉及匿名場景）。
  */
-export function CommentInput({ reactionType, onSubmit, disabled, className }: CommentInputProps) {
+export function CommentInput({
+  reactionType,
+  onSubmit,
+  disabled,
+  className,
+  participants = [],
+}: CommentInputProps) {
   const [value, setValue] = useState("");
   const textareaRef = useRef<HTMLTextAreaElement>(null);
   const prevReactionTypeRef = useRef<ReactionTypeType | null | undefined>(undefined);
+  const { handleMentionSelect, getActiveMentionIds, reset: resetMentions } = useMentionInput();
 
   // reactionType 改變時聚焦輸入框
   useEffect(() => {
@@ -52,9 +71,10 @@ export function CommentInput({ reactionType, onSubmit, disabled, className }: Co
   const handleSubmit = useCallback(() => {
     const trimmed = value.trim();
     if (!trimmed) return;
-    onSubmit(trimmed, reactionType);
+    onSubmit(trimmed, reactionType, getActiveMentionIds(trimmed));
     setValue("");
-  }, [value, reactionType, onSubmit]);
+    resetMentions();
+  }, [value, reactionType, onSubmit, getActiveMentionIds, resetMentions]);
 
   const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
     if (e.key === "Enter" && !e.shiftKey && !e.nativeEvent.isComposing) {
@@ -65,14 +85,16 @@ export function CommentInput({ reactionType, onSubmit, disabled, className }: Co
 
   return (
     <div className={cn("flex gap-2 items-center", className)}>
-      <textarea
-        ref={textareaRef}
+      <MentionInput
+        inputRef={textareaRef}
         value={value}
-        onChange={(e) => setValue(e.target.value)}
+        onChange={setValue}
         onKeyDown={handleKeyDown}
         placeholder={placeholder}
         rows={1}
+        participants={participants}
         disabled={disabled}
+        onMentionSelect={handleMentionSelect}
         className="flex-1 resize-none rounded-lg border border-[#E4EAE9] bg-white px-4 py-2 text-sm text-[#295E5C] placeholder:text-[#9FB5B8] focus:outline-none focus:border-logo-cyan transition-colors h-10 disabled:opacity-50"
       />
       <Button
