+++ packages/features/mention/src/index.ts
@@ -1,0 +1,7 @@
+// Components
+
+export type { MentionInputProps } from "./components/mention-input";
+export { MentionInput } from "./components/mention-input";
+export type { MentionCandidate } from "./hooks/use-mention-input";
+// Hooks
+export { useMentionInput } from "./hooks/use-mention-input";
