+++ packages/features/mention/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@daodao/config/react.json",
+  "compilerOptions": {
+    "outDir": "./dist"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
